refactor(git-utils) to follow coding conventions
