--- packages/amplication-git-utils/src/Dto/enums/EnumGitOrganizationType.ts
@@ -1,10 +1,0 @@
-import { registerEnumType } from "@nestjs/graphql";
-
-export enum EnumGitOrganizationType {
-  User = "User",
-  Organization = "Organization",
-}
-
-registerEnumType(EnumGitOrganizationType, {
-  name: "EnumGitOrganizationType",
-});
