--- packages/amplication-git-utils/src/Dto/enums/EnumGitProvider.ts
@@ -1,9 +1,0 @@
-import { registerEnumType } from "@nestjs/graphql";
-
-export enum EnumGitProvider {
-  Github = "Github",
-}
-
-registerEnumType(EnumGitProvider, {
-  name: "EnumGitProvider",
-});
