--- packages/amplication-git-utils/src/__mocks__/Constants.ts
@@ -1,2 +1,0 @@
-export const INSTALLATION_URL = "ExampleInstallationUrl";
-export const PR_HTML_URL = "ExamplePrHtmlUrl";
