--- packages/amplication-git-utils/src/__mocks__/GithubFile.ts
@@ -1,8 +1,0 @@
-import { GithubFile } from "../Dto/entities/GithubFile";
-
-export const GIT_HUB_FILE: GithubFile = {
-  name: "exampleGithubFileName",
-  path: "examplePath",
-  content: "exampleContent",
-  htmlUrl: "exampleHtmlUrl",
-};
