--- packages/amplication-git-utils/src/__mocks__/RemoteGitRepository.ts
@@ -1,9 +1,0 @@
-import { RemoteGitRepository } from "../Dto/entities/RemoteGitRepository";
-
-export const TEST_GIT_REPO: RemoteGitRepository = {
-  admin: true,
-  fullName: "tupe12334/repo",
-  name: "repo",
-  private: false,
-  url: "localhost/repo",
-};
