--- packages/amplication-git-utils/src/contracts/GitResourceMeta.ts
@@ -1,8 +1,0 @@
-import { IsString } from "class-validator";
-
-export class GitResourceMeta {
-  @IsString()
-  serverPath: string;
-  @IsString()
-  adminUIPath: string;
-}
