+++ packages/amplication-git-utils/src/git/dto/remote-git-organization.dto.ts
@@ -1,5 +1,5 @@
 import { Field, ObjectType } from "@nestjs/graphql";
-import { EnumGitOrganizationType } from "../enums/EnumGitOrganizationType";
+import { EnumGitOrganizationType } from "../git.types";
 
 @ObjectType({
   isAbstract: true,
