+++ packages/amplication-git-utils/src/git/dto/remote-git-repository.ts
