+++ packages/amplication-git-utils/src/git/git-service-factory.ts
@@ -1,13 +1,12 @@
 import { Injectable } from "@nestjs/common";
-import { IGitClient } from "../contracts/IGitClient";
-import { EnumGitProvider } from "../Dto/enums/EnumGitProvider";
-import { GithubService } from "../providers/github.service";
-import { INVALID_SOURCE_CONTROL_ERROR_MESSAGE } from "./constants";
+import { GithubService } from "./github.service";
+import { INVALID_SOURCE_CONTROL_ERROR_MESSAGE } from "./git.constants";
+import { EnumGitProvider, GitClient } from "./git.types";
 
 @Injectable()
 export class GitServiceFactory {
   constructor(protected githubService: GithubService) {}
-  getService(gitProvider: EnumGitProvider): IGitClient {
+  getService(gitProvider: EnumGitProvider): GitClient {
     switch (gitProvider) {
       case EnumGitProvider.Github:
         return this.githubService;
