+++ packages/amplication-git-utils/src/git/git.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+import { GitService } from "./git.service";
+import { GithubService } from "./github.service";
+import { GitServiceFactory } from "./git-service-factory";
+
+@Module({
+  providers: [GitService, GithubService, GitServiceFactory],
+  exports: [GitService, GitServiceFactory, GithubService],
+})
+export class GitModule {}
