+++ packages/amplication-git-utils/src/index.ts
@@ -1,6 +1,5 @@
-export { GitModule } from "./git.module";
-export { GitService } from "./services/git.service";
-export { GithubService } from "./providers/github.service";
-export { GitServiceFactory } from "./utils/GitServiceFactory";
-export { EnumGitOrganizationType } from "./Dto/enums/EnumGitOrganizationType";
-export { EnumGitProvider } from "./Dto/enums/EnumGitProvider";
+export { GitModule } from "./git/git.module";
+export { GitService } from "./git/git.service";
+export { GithubService } from "./git/github.service";
+export { GitServiceFactory } from "./git/git-service-factory";
+export { EnumGitOrganizationType, EnumGitProvider } from "./git/git.types";
