--- packages/amplication-git-utils/src/types.ts
@@ -1,4 +1,0 @@
-export type PrModule = {
-  path: string;
-  code: string | null;
-};
