+++ packages/amplication-git-utils/src/utils/amplication-ignore-manger.ts
