+++ packages/amplication-git-utils/tests/git/git.service.spec.ts
@@ -1,14 +1,12 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { EnumGitProvider } from "../../Dto/enums/EnumGitProvider";
-import { GitService } from "../../services/git.service";
-import { GitServiceFactory } from "../../utils/GitServiceFactory";
-import { MOCK_GIT_SERVICE_FACTORY } from "../../__mocks__/GitServiceFactory.mock";
-import { EnumGitOrganizationType } from "../../Dto/enums/EnumGitOrganizationType";
-import { TEST_GIT_REPO } from "../../__mocks__/RemoteGitRepository";
-import { TEST_GIT_REPOS } from "../../__mocks__/RemoteGitRepositories";
-import { TEST_GIT_REMOTE_ORGANIZATION } from "../../__mocks__/RemoteGitOrganization";
-import { INSTALLATION_URL, PR_HTML_URL } from "../../__mocks__/Constants";
-import { GIT_HUB_FILE } from "../../__mocks__/GithubFile";
+import {
+  EnumGitOrganizationType,
+  EnumGitProvider,
+} from "../../src/git/git.types";
+import { GitService } from "../../src/git/git.service";
+import { GitServiceFactory } from "../../src/git/git-service-factory";
+import { MOCK_GIT_SERVICE_FACTORY } from "./mocks";
+import { GIT_HUB_FILE, INSTALLATION_URL, PR_HTML_URL, TEST_GIT_REMOTE_ORGANIZATION, TEST_GIT_REPO, TEST_GIT_REPOS } from "./git.constants";
 
 describe("GitService", () => {
   let gitService: GitService;
