+++ packages/amplication-git-utils/tests/git/mocks.ts
@@ -1,10 +1,18 @@
-import { GithubService } from "../providers/github.service";
+import { EnumGitProvider } from "../../src/git/git.types";
+import { GitServiceFactory } from "../../src/git/git-service-factory";
 import { mock } from "jest-mock-extended";
-import { TEST_GIT_REPOS } from "./RemoteGitRepositories";
-import { TEST_GIT_REPO } from "./RemoteGitRepository";
-import { TEST_GIT_REMOTE_ORGANIZATION } from "./RemoteGitOrganization";
-import { INSTALLATION_URL, PR_HTML_URL } from "./Constants";
-import { GIT_HUB_FILE } from "./GithubFile";
+
+import {
+  GIT_HUB_FILE,
+  INSTALLATION_URL,
+  PR_HTML_URL,
+  TEST_GIT_REMOTE_ORGANIZATION,
+  TEST_GIT_REPO,
+  TEST_GIT_REPOS,
+} from "./git.constants";
+import { GithubService } from "../../src/git/github.service";
+
+export const MOCK_GIT_SERVICE_FACTORY = mock<GitServiceFactory>();
 
 export const MOCK_GITHUB_SERVICE = mock<GithubService>({});
 MOCK_GITHUB_SERVICE.getOrganizationRepos.mockReturnValue(
@@ -32,3 +40,7 @@ MOCK_GITHUB_SERVICE.getFile.mockReturnValue(Promise.resolve(GIT_HUB_FILE));
 MOCK_GITHUB_SERVICE.createPullRequest.mockReturnValue(
   Promise.resolve(PR_HTML_URL)
 );
+
+MOCK_GIT_SERVICE_FACTORY.getService
+  .calledWith(EnumGitProvider.Github)
+  .mockReturnValue(MOCK_GITHUB_SERVICE);
